Small changes to wobba/supportfolderinopenxml
