+++ Core/OfficeDevPnP.Core/Framework/Provisioning/Connectors/OpenXML/PnPPackage.cs
@@ -13,7 +13,7 @@ namespace OfficeDevPnP.Core.Framework.Provisioning.Connectors.OpenXML
     /// <summary>
     /// Defines a PnP OpenXML package file
     /// </summary>
-    public class PnPPackage : IDisposable
+    public partial class PnPPackage : IDisposable
     {
         #region Constant strings
 
@@ -191,16 +191,6 @@ namespace OfficeDevPnP.Core.Framework.Provisioning.Connectors.OpenXML
         {
             fileName = fileName.TrimStart('/');
             string uriStr = U_DIR_FILES + fileName;
-            PackagePart part = CreatePackagePart(R_PROVISIONINGTEMPLATE_FILE, CT_FILE, uriStr, FilesOriginPart);
-            SetPackagePartValue(value, part);
-        }
-
-        [Obsolete("Folders are now handled inside the mapping file")]
-        public void AddFile(string fileName, string folder, Byte[] value)
-        {
-            fileName = fileName.TrimStart('/');
-            folder = !String.IsNullOrEmpty(folder) ? (folder.TrimStart('/').TrimEnd('/') + "/") : String.Empty;
-            string uriStr = U_DIR_FILES + folder + fileName;
             PackagePart part = CreatePackagePart(R_PROVISIONINGTEMPLATE_FILE, CT_FILE, uriStr, FilesOriginPart);
             SetPackagePartValue(value, part);
         }
