+++ Core/OfficeDevPnP.Core/OfficeDevPnP.Core.csproj
@@ -443,6 +443,7 @@
     <Compile Include="Framework\Provisioning\Connectors\OpenXML\PnPPackageFormatException.cs" />
     <Compile Include="Framework\Provisioning\Connectors\OpenXML\PnPPackageHelper.cs" />
     <Compile Include="Framework\Provisioning\Connectors\SharePointConnector.cs" />
+    <Compile Include="Framework\Provisioning\Deprecated\Connectors\OpenXML\PnPPackage.cs" />
     <Compile Include="Framework\Provisioning\Deprecated\Extensibility\ExtensibilityManager.deprecated.cs" />
     <Compile Include="Framework\Provisioning\Deprecated\Extensibility\IProvisioningExtensibilityProvider.deprecated.cs" />
     <Compile Include="Framework\Provisioning\Deprecated\Model\ComposedLook.deprecated.cs" />
